+++ lib/src/cli/commands/commands.dart
@@ -1,4 +1,5 @@
 export 'cache_command.dart';
+export 'collect_command.dart';
 export 'config_command.dart';
 export 'crawl_command.dart';
 export 'cues_command.dart';